+++ src/main/java/com/mycomp/client/TableService.java
@@ -12,6 +12,6 @@ import java.util.List;
 @RemoteServiceRelativePath("TableService")
 public interface TableService extends RemoteService {
     List<UsersEntity> getAll(int firstId, int lastId);
-    Integer getRowCount();
+    Long getRowCount();
 
 }
